Load test (try 5).
